requirements.txt
azure-functions==1.17.0
pandas==1.4.3
pydantic[dotenv]>=1.8.0,<2.0.0
azure-storage-blob==12.14.1
python-keycloak==0.27.0
pre-commit==2.19.0
asyncpg==0.27.0
azure-functions-durable==1.2.8
pyarrow==11.0.0
